OV-748: Stage 3 of refactoring sent emails. Move all classes into notifications folder.
