+++ src/main/kotlin/uk/gov/justice/digital/hmpps/officialvisitsapi/config/GovNotifyConfiguration.kt
@@ -4,13 +4,13 @@ import org.slf4j.LoggerFactory
 import org.springframework.beans.factory.annotation.Value
 import org.springframework.context.annotation.Bean
 import org.springframework.context.annotation.Configuration
-import uk.gov.justice.digital.hmpps.officialvisitsapi.service.emails.EmailService
-import uk.gov.justice.digital.hmpps.officialvisitsapi.service.emails.EmailTemplate
-import uk.gov.justice.digital.hmpps.officialvisitsapi.service.emails.EmailTemplates
-import uk.gov.justice.digital.hmpps.officialvisitsapi.service.emails.EmailType.OFFICIAL_VISIT_CANCELLED
-import uk.gov.justice.digital.hmpps.officialvisitsapi.service.emails.EmailType.OFFICIAL_VISIT_CREATED
-import uk.gov.justice.digital.hmpps.officialvisitsapi.service.emails.EmailType.OFFICIAL_VISIT_UPDATED
-import uk.gov.justice.digital.hmpps.officialvisitsapi.service.emails.GovNotifyEmailService
+import uk.gov.justice.digital.hmpps.officialvisitsapi.service.notifications.EmailService
+import uk.gov.justice.digital.hmpps.officialvisitsapi.service.notifications.EmailTemplate
+import uk.gov.justice.digital.hmpps.officialvisitsapi.service.notifications.EmailTemplates
+import uk.gov.justice.digital.hmpps.officialvisitsapi.service.notifications.EmailType.OFFICIAL_VISIT_CANCELLED
+import uk.gov.justice.digital.hmpps.officialvisitsapi.service.notifications.EmailType.OFFICIAL_VISIT_CREATED
+import uk.gov.justice.digital.hmpps.officialvisitsapi.service.notifications.EmailType.OFFICIAL_VISIT_UPDATED
+import uk.gov.justice.digital.hmpps.officialvisitsapi.service.notifications.GovNotifyEmailService
 import uk.gov.service.notify.NotificationClient
 import java.util.UUID
