+++ src/main/kotlin/uk/gov/justice/digital/hmpps/officialvisitsapi/service/SentEmailsService.kt
@@ -10,10 +10,10 @@ import uk.gov.justice.digital.hmpps.officialvisitsapi.entity.SentNotificationEnt
 import uk.gov.justice.digital.hmpps.officialvisitsapi.model.request.NotificationSearchRequest
 import uk.gov.justice.digital.hmpps.officialvisitsapi.model.response.SentNotification
 import uk.gov.justice.digital.hmpps.officialvisitsapi.repository.NotificationSearchRepository
-import uk.gov.justice.digital.hmpps.officialvisitsapi.service.emails.EmailType
 import uk.gov.justice.digital.hmpps.officialvisitsapi.service.metrics.MetricsEvents
 import uk.gov.justice.digital.hmpps.officialvisitsapi.service.metrics.MetricsService
 import uk.gov.justice.digital.hmpps.officialvisitsapi.service.metrics.NotificationSearchInfo
+import uk.gov.justice.digital.hmpps.officialvisitsapi.service.notifications.EmailType
 import java.time.format.DateTimeFormatter
 
 @Service
