+++ src/main/kotlin/uk/gov/justice/digital/hmpps/officialvisitsapi/service/notifications/EmailService.kt
@@ -1,4 +1,4 @@
-package uk.gov.justice.digital.hmpps.officialvisitsapi.service.emails
+package uk.gov.justice.digital.hmpps.officialvisitsapi.service.notifications
 
 import uk.gov.justice.digital.hmpps.officialvisitsapi.common.toHourMinuteStyle
 import uk.gov.justice.digital.hmpps.officialvisitsapi.common.toMediumFormatStyle
